Nav: Simplify mobile drawer
The Nav drawer had dead code in the form of a drawer that was supposed to display on medium-size and up browsers permanently open. That permanent drawer has been replaced by the Nav avatar and is no longer rendered. This means the dead code was bloating the component and confusing the control flow. It has been removed so that the Nav component is only used to display the mobile nav drawer.